backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
networkx
numpy
scikit-fuzzy
torch
torch-geometric
pygad